requirements.txt
discord.py
pillow
pandas
xlsxwriter
python-dotenv
